feat: Allow configurable test size
Introduces a new `testSize` setting persisted via localStorage. This allows users to adjust the number of words in daily tests and mistake reviews. The `wordService` has been updated to accept this configurable size, dynamically calculating the proportion of revision words based on the new target size.

Additionally, the home screen now displays and allows modification of this `testSize` setting.